Move v-short directive to plugins directory
