+++ client/src/components/History/CurrentCollection/CollectionDetails.vue
@@ -18,10 +18,10 @@
 </template>
 
 <script>
-import short from "components/directives/v-short";
-import DetailsLayout from "components/History/Layout/DetailsLayout";
-import CollectionDescription from "components/History/Content/Collection/CollectionDescription";
-import { JobStateSummary } from "components/History/Content/Collection/JobStateSummary";
+import short from "@/components/plugins/short.js";
+import DetailsLayout from "@/components/History/Layout/DetailsLayout.vue";
+import CollectionDescription from "@/components/History/Content/Collection/CollectionDescription.vue";
+import { JobStateSummary } from "@/components/History/Content/Collection/JobStateSummary.js";
 
 export default {
     components: {
